moving files from mock debugger to adapter dir
