+++ src/daffodilDebugger.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs'
 import * as unzip from 'unzip-stream'
 import * as os from 'os'
 import * as child_process from 'child_process'
-import { deactivate } from './extension'
+import { deactivate } from './adapter/extension'
 import { LIB_VERSION } from './version'
 import { HttpClient } from 'typed-rest-client/HttpClient'
 import XDGAppPaths from 'xdg-app-paths'
